+++ great_expectations/execution_engine/exceptions.py
@@ -1,0 +1,5 @@
+from great_expectations.exceptions import GreatExpectationsError
+
+
+class ExecutionEngineError(GreatExpectationsError):
+    pass